friendry_review.pyの関数化
